requirements.txt
django
djangorestframework
django-redis
mysqlclient
PyJWT
rsa
bcrypt